precompile level set
